first-time use user-setup completed
